Moving link-related code into LinkBuilder classes
