+++ spec/presenters/quadicons/link_builders.rb
@@ -1,0 +1,7 @@
+module Quadicons
+  module LinkBuilders
+    def self.builder_for(quadicon)
+      # find link builder based on name
+    end
+  end
+end
