--- spec/presenters/quadicons/url_builder_spec.rb
@@ -1,3 +1,0 @@
-describe Quadicons::UrlBuilder, :type => :helper do
-  subject(:url) { Quadicons::UrlBuilder.new(record, helper).url }
-end
